LPS-131186 Reusing getDescription
